// src/app/crytos/interfaces/crytos.interfaces.ts
export interface Crytos {
    symbol: string;
    priceChange: number;
    priceChangePercent: number;
    weightedAvgPrice: number;
    prevClosePrice: number;
    lastPrice: number;
    lastQty: number;
    bidPrice: number;
    bidQty: number;
    askPrice: number;
    askQty: number;
    openPrice: number;
    highPrice: number;
    lowPrice: number;
    volume: number;
    quoteVolume: number;
    openTime: number;
    closeTime: number;
    firstId: number;
    lastId: number;
    count: number;
}
export interface CrytosCard {
    symbol: string;
    priceChange: number;
    priceChangePercent: number;
    lastPrice: number;
    highPrice: number;
    lowPrice: number;
    volume: number;
}

export interface Coin {
    symbol: string;
  
    
}